Rename Home to About
PDJ-21

Took 6 minutes
