+++ PlaylistDJ.Frontend/src/pages/index.ts
@@ -1,4 +1,4 @@
-import Home from './Home.svelte'
+import About from './About.svelte'
 import HomeLoggedIn from './HomeLoggedIn.svelte'
 import NotFound from './NotFound.svelte'
 import Playlist from './Playlist.svelte'
@@ -6,4 +6,4 @@ import PlaylistList from './PlaylistList.svelte'
 import Settings from './Settings.svelte'
 import Social from './Social.svelte'
 
-export { Home, HomeLoggedIn, NotFound, Playlist, PlaylistList, Settings, Social }
+export { About, HomeLoggedIn, NotFound, Playlist, PlaylistList, Settings, Social }
